requirements.txt
appdirs==1.4.3
asgiref==3.2.3
attrs==19.1.0
black==19.3b0
Click==7.0
defusedxml==0.6.0
diff-match-patch==20181111
dj-database-url==0.5.0
Django==3.0.4
django-admin-charts==0.19.0
django-bower==5.2.0
django-classy-tags==1.0.0
django-formtools==2.2
django-heroku==0.3.1
django-import-export==2.0.2
django-jsonfield==1.4.0
django-nested-inline==0.4.2
django-nvd3==0.9.7
django-sekizai==1.1.0
django-treebeard==4.3.1
djangocms-admin-style==1.5.0
djangocms-installer==1.2.1
djcacheutils==3.0.0
entrypoints==0.3
et-xmlfile==1.0.1
flake8==3.7.8
gunicorn==20.0.4
jdcal==1.4.1
Jinja2==2.11.2
MarkupPy==1.14
MarkupSafe==1.1.1
mccabe==0.6.1
odfpy==1.4.1
openpyxl==3.0.3
psycopg2==2.8.4
pycodestyle==2.5.0
pyflakes==2.1.1
python-dateutil==2.8.1
python-memcached==1.59
python-nvd3==0.14.2
python-slugify==1.1.4
pytz==2019.3
PyYAML==5.3
six==1.14.0
sqlparse==0.3.1
tablib==1.1.0
toml==0.10.0
tzlocal==2.0.0
Unidecode==1.1.1
whitenoise==5.0.1
xlrd==1.2.0
xlwt==1.3.0
isoweek==1.3.3
django-sb-admin==5.5.1.1
pandas==1.0.5
NumPy==1.18.0
pdfkit==0.6.1